--- a/conda_build/os_utils/macho.py
+++ b/conda_build/os_utils/macho.py
@@ -224,7 +224,7 @@ def delete_rpath(path, rpath, verbose=False):
         % p.returncode)
 
 
-def install_name_change(path, cb_func, verbose=False):
+def install_name_change(path, cb_func, dylibs, verbose=False):
     """Change dynamic shared library load name or id name of Mach-O Binary `path`.
 
     `cb_func` is called for each shared library load command. The dictionary of
@@ -234,7 +234,7 @@ def install_name_change(path, cb_func, verbose=False):
     When dealing with id load commands, `install_name_tool -id` is used.
     When dealing with dylib load commands `install_name_tool -change` is used.
     """
-    dylibs = otool(path)
+
     changes = []
     for index, dylib in enumerate(dylibs):
         new_name = cb_func(path, dylib)