@msarahan 2017-12-12
great start!  How are you getting your line profile info?  I'd like to combine any instructions you have with my instructions for doing the pytest-profiling stuff.

I'll take an in-depth look at this in a little while, but I'm very impressed with what you have so far.  I agree with your hunch that you should be able to assume that the prefix doesn't change.  If it were changing at this point, there would be much bigger problems.