@rlizzo 2017-12-12
That's an interesting idea. Looking forward to testing it out myself tonight. I was thinking more of a stateful dictionary, with files that we've already seen as the key, the the value a bool of weather to include them in the files list. 

Since these functions aren't mac OS specific, it might be cleaner to split this off into it's own methods? That or start passing in some kwargs to change function behavior.   Not entirely sure what the most pythonic way is  yet. Need to put some more thought into it.

l'll send you the profiling code when I can get to a computer later. Currently dictating this in the car.