@msarahan 2017-12-12
This latest idea is from @mingwandroid.  I think it's the best approach.  It avoids any complexity around caching not working right, and achieves the same purpose.  Please give it a spin tonight and let us know how it goes.