@rlizzo 2017-12-12
Will do. Thanks

On Dec 12, 2017 3:10 PM, "Mike Sarahan" <notifications@github.com> wrote:

> This latest idea is from @mingwandroid <https://github.com/mingwandroid>.
> I think it's the best approach. It avoids any complexity around caching not
> working right, and achieves the same purpose. Please give it a spin tonight
> and let us know how it goes.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/conda/conda-build/pull/2564#issuecomment-351179113>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AQi945mxVL2Vbjesu-sijjbxHii6LWP4ks5s_t2sgaJpZM4Q-nrg>
> .
>